Move root resources from `service/controller/clusterapi`
and `.../clusterapi/v22` to `service/controller`, merging `error.go`
from both packages.

Also make `ClusterResourceSetConfig` and `MachineDeploymentResourceSetConfig`
private as they are only used in the same package now.
